+++ model/storage-private/src/main/java/org/keycloak/services/scheduled/OpenIdFederationIdPExpirationTask.java
@@ -4,6 +4,7 @@ import org.jboss.logging.Logger;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
+import org.keycloak.protocol.trustchain.TrustChainProcessor;
 import org.keycloak.timer.ScheduledTask;
 import org.keycloak.timer.TimerProvider;
 
@@ -32,11 +33,10 @@ public class OpenIdFederationIdPExpirationTask implements ScheduledTask {
         IdentityProviderModel idp = session.identityProviders().getByAlias(alias);
         if (idp == null || !"openid-federation".equals(idp.getProviderId()) ) {
             TimerProvider timer = session.getProvider(TimerProvider.class);
-            timer.cancelTask("OpenIdFederationIdPExpirationTask_" + alias);
+            timer.cancelTaskAndNotify("OpenIdFederationIdPExpirationTask_" + alias);
         } else {
-            //TODO retry for updating IdP
-            idp.setEnabled(false);
-            session.identityProviders().update(idp);
+            TrustChainProcessor trustChainProcessor = session.getProvider(TrustChainProcessor.class, "openid-federation");
+            trustChainProcessor.updateIdP(idp, realm);
         }
     }
 }