+++ services/src/main/java/org/keycloak/protocol/oidc/endpoints/AuthorizationEndpoint.java
@@ -75,7 +75,6 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.core.Response;
-import org.keycloak.utils.OpenIdFederationTrustChainProcessor;
 import org.keycloak.utils.OpenIdFederationUtils;
 
 import java.util.List;
@@ -114,13 +113,11 @@ public class AuthorizationEndpoint extends AuthorizationEndpointBase {
 
     private AuthorizationEndpointRequest request;
     private String redirectUri;
-    private OpenIdFederationTrustChainProcessor trustChainProcessor;
     private boolean automaticRegistration = false;
 
     public AuthorizationEndpoint(KeycloakSession session, EventBuilder event) {
         super(session, event);
         event.event(EventType.LOGIN);
-        this.trustChainProcessor = new OpenIdFederationTrustChainProcessor(session);
     }
 
     @POST
@@ -297,24 +294,9 @@ public class AuthorizationEndpoint extends AuthorizationEndpointBase {
         client = realm.getClientByClientId(clientId);
         if (UriUtils.isUri(clientId) && (client == null || !client.isEnabled()) && realm.isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum.OPENID_PROVIDER, ClientRegistrationTypeEnum.AUTOMATIC)) {
                 try {
-                    String rpMetadata = OpenIdFederationUtils.getSelfSignedToken(clientId, session);
-                    EntityStatement rpEntityStatement = trustChainProcessor.parseAndValidateSelfSigned(rpMetadata);
-                    trustChainProcessor.validationRules(rpEntityStatement, false);
-                    logger.info("starting validating trust chains");
-                    TrustChainResolution validChain = trustChainProcessor.constructTrustChains(rpEntityStatement, realm.getOpenIdFederationsTrustAnchors(), true);
-
-                    if (validChain == null) {
-                        throw new ErrorResponseException(Errors.INVALID_TRUST_ANCHOR, "No trusted trust anchor could be found", Response.Status.NOT_FOUND);
-                    }
-
-                    EventBuilder event = new EventBuilder(session.getContext().getRealm(), session, session.getContext().getConnection());
-                    OpenIdFederationClientRegistrationService provider = new OpenIdFederationClientRegistrationService(session);
-                    provider.setEvent(event);
-                    provider.setAuth(new ClientRegistrationAuth(session, provider, event, "openid-connect"));
-                    ClientRepresentation clientRepresentation = provider.createOrUpdateClient(rpEntityStatement, (RPMetadata) validChain.getMetadataAfterPolicies());
-                    client = realm.getClientByClientId(clientRepresentation.getClientId());
+                    client = OpenIdFederationUtils.createOrUpdateAutomaticClient(clientId, session, realm);
                     automaticRegistration = true;
-                } catch (ErrorPageException e) {
+                } catch (ErrorPageException | ErrorResponseException e) {
                     throw e;
                 } catch (Exception e) {
                     throw new ErrorResponseException(Errors.INVALID_REQUEST, e.getMessage(), Response.Status.BAD_REQUEST);