+++ services/src/main/java/org/keycloak/protocol/oidc/federation/OpenIdFederationAuthzEndpointRequestObjectParser.java
@@ -29,9 +29,10 @@ public class OpenIdFederationAuthzEndpointRequestObjectParser extends AuthzEndpo
             throw new RuntimeException("The request_uri claim should not be set in the request object");
         }
 
-        if (requestParams.has("sub")  || ! requestParams.has("jti")
-                || !requestParams.has("exp") || !requestParams.has("iss")
-                ||  !Urls.realmIssuer(session.getContext().getUri(UrlType.FRONTEND).getBaseUri(), session.getContext().getRealm().getName()).equals(requestParams.get("aud"))) {
+        if (requestParams.has("sub")  || ! requestParams.has("jti") || !requestParams.has("exp")
+                || !requestParams.has("client_id") || !requestParams.has("iss") || !requestParams.has("aud")
+                || !client.getClientId().equals(requestParams.get("client_id").asText()) || !client.getClientId().equals(requestParams.get("iss").asText())
+                || !Urls.realmIssuer(session.getContext().getUri(UrlType.FRONTEND).getBaseUri(), session.getContext().getRealm().getName()).equals(requestParams.get("aud").asText())) {
             throw new RuntimeException(Messages.OPENID_FEDERATION_AUTOMATIC_FALSE_REQUEST_OBJECT);
         }