+++ services/src/main/java/org/keycloak/services/clientregistration/policy/impl/ClientScopesClientRegistrationPolicy.java
@@ -59,9 +59,26 @@ public class ClientScopesClientRegistrationPolicy implements ClientRegistrationP
         allowedScopeNames.addAll(getAllowedScopeNames(realm, true));
         allowedScopeNames.addAll(getAllowedScopeNames(realm, false));
 
-
         checkClientScopesAllowed(requestedDefaultScopeNames, allowedScopeNames);
         checkClientScopesAllowed(requestedOptionalScopeNames, allowedScopeNames);
+
+        if (componentModel.get(ClientScopesClientRegistrationPolicyFactory.ALLOW_DEFAULT_SCOPES, true) && (requestedDefaultScopeNames != null || requestedOptionalScopeNames != null)) {
+            //if requested scopes list is not empty and add-default-scopes is true, add realm default scopes
+            List<String> defaultRealmScopes = realm.getDefaultClientScopesStream(true).map(ClientScopeModel::getName).collect(Collectors.toList());
+            if (!defaultRealmScopes.isEmpty()) {
+                if (context.getClient().getDefaultClientScopes() == null) {
+                    context.getClient().setDefaultClientScopes(new ArrayList<>());
+                }
+                for (String s : defaultRealmScopes) {
+                    if (!context.getClient().getDefaultClientScopes().contains(s)) {
+                        context.getClient().getDefaultClientScopes().add(s);
+                    }
+                }
+                if (requestedOptionalScopeNames != null) {
+                    context.getClient().getOptionalClientScopes().removeIf(x -> defaultRealmScopes.contains(x));
+                }
+            }
+        }
     }
 
     @Override