+++ services/src/main/java/org/keycloak/services/resources/admin/IdentityProvidersResource.java
@@ -16,10 +16,6 @@
  */
 
 package org.keycloak.services.resources.admin;
-
-import jakarta.ws.rs.NotFoundException;
-import jakarta.ws.rs.core.UriInfo;
-import org.apache.http.entity.StringEntity;
 import org.eclipse.microprofile.openapi.annotations.Operation;
 import org.eclipse.microprofile.openapi.annotations.extensions.Extension;
 import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
@@ -27,9 +23,7 @@ import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.resteasy.reactive.NoCache;
-import org.keycloak.broker.oidc.OAuth2IdentityProviderConfig;
 import org.keycloak.broker.oidc.OIDCIdentityProviderConfig;
-import org.keycloak.broker.oidc.OIDCIdentityProviderFactory;
 import org.keycloak.broker.oidc.federation.OpenIdFederationIdentityProviderConfig;
 import org.keycloak.broker.oidc.federation.OpenIdFederationIdentityProviderFactory;
 import org.keycloak.broker.provider.IdentityProviderFactory;
@@ -39,14 +33,9 @@ import org.keycloak.connections.httpclient.HttpClientProvider;
 import org.keycloak.events.admin.OperationType;
 import org.keycloak.events.admin.ResourceType;
 import org.keycloak.http.FormPartValue;
-import org.keycloak.http.simple.SimpleHttp;
-import org.keycloak.http.simple.SimpleHttpResponse;
-import org.keycloak.jose.jwk.JSONWebKeySet;
 import org.keycloak.models.IdentityProviderModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.ModelDuplicateException;
-import org.keycloak.models.OpenIdFederationConfig;
-import org.keycloak.models.OpenIdFederationGeneralConfig;
 import org.keycloak.models.RealmModel;
 import org.keycloak.models.enums.ClientRegistrationTypeEnum;
 import org.keycloak.models.enums.EntityTypeEnum;
@@ -54,30 +43,17 @@ import org.keycloak.models.utils.ModelToRepresentation;
 import org.keycloak.models.utils.RepresentationToModel;
 import org.keycloak.models.utils.StripSecretsUtils;
 import org.keycloak.protocol.oidc.OIDCConfigAttributes;
-import org.keycloak.protocol.oidc.OIDCLoginProtocol;
-import org.keycloak.protocol.oidc.OIDCLoginProtocolService;
-import org.keycloak.protocol.oidc.OIDCWellKnownProvider;
+import org.keycloak.protocol.trustchain.TrustChainProcessor;
 import org.keycloak.utils.OpenIdFederationTrustChainProcessorFactory;
 import org.keycloak.representations.idm.IdentityProviderRepresentation;
-import org.keycloak.representations.openid_federation.EntityStatement;
-import org.keycloak.representations.openid_federation.EntityStatementExplicitResponse;
-import org.keycloak.representations.openid_federation.Metadata;
-import org.keycloak.representations.openid_federation.OPMetadata;
-import org.keycloak.representations.openid_federation.RPMetadata;
-import org.keycloak.representations.openid_federation.TrustChainResolution;
 import org.keycloak.services.ErrorResponse;
-import org.keycloak.services.Urls;
 import org.keycloak.services.resources.KeycloakOpenAPI;
-import org.keycloak.services.resources.RealmsResource;
 import org.keycloak.services.resources.admin.fgap.AdminPermissionEvaluator;
 import org.keycloak.services.scheduled.AutoUpdateIdentityProviders;
 import org.keycloak.services.scheduled.ClusterAwareScheduledTaskRunner;
 import org.keycloak.services.scheduled.OpenIdFederationIdPExpirationTask;
 import org.keycloak.services.util.ResourcesUtil;
 import org.keycloak.timer.TimerProvider;
-import org.keycloak.urls.UrlType;
-import org.keycloak.util.TokenUtil;
-import org.keycloak.utils.OpenIdFederationUtils;
 import org.keycloak.utils.ReservedCharValidator;
 
 import jakarta.ws.rs.BadRequestException;
@@ -92,19 +68,12 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.MultivaluedMap;
 import jakarta.ws.rs.core.Response;
 import org.keycloak.utils.StringUtil;
-import org.keycloak.utils.TrustChainProcessor;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static jakarta.ws.rs.core.Response.Status.BAD_REQUEST;
@@ -307,54 +276,14 @@ public class IdentityProvidersResource {
     private IdentityProviderModel createModelForOpenIdFederation(IdentityProviderRepresentation representation){
         if (realm.isOpenIdFederationTypeRegistrationSupported(EntityTypeEnum.OPENID_RELYING_PARTY, ClientRegistrationTypeEnum.EXPLICIT) && representation.getConfig().get(OIDCIdentityProviderConfig.ISSUER) != null) {
             try {
-                OpenIdFederationGeneralConfig federationGeneralConfig = realm.getOpenIdFederationGeneralConfig();
-                OpenIdFederationConfig federationConfig = federationGeneralConfig.getOpenIdFederationList().stream().filter(x -> representation.getConfig().get(OpenIdFederationIdentityProviderConfig.TRUST_ANCHOR_ID).equals(x.getTrustAnchor())).findAny().orElseThrow(() -> new NotFoundException("Trust anchor does not exist"));
-                TrustChainProcessor trustChainProcessor = session.getProvider(TrustChainProcessor.class, OpenIdFederationTrustChainProcessorFactory.PROVIDER_ID);
-                String opIssuer = representation.getConfig().get(OIDCIdentityProviderConfig.ISSUER);
-                EntityStatement opStatement = trustChainProcessor.parseAndValidateSelfSigned(OpenIdFederationUtils.getSelfSignedToken(opIssuer, session));
-                if (!trustChainProcessor.validateEntityStatementFields(opStatement, opIssuer, opIssuer) || opStatement.getMetadata().getOpenIdProviderMetadata() == null || !opStatement.getMetadata().getOpenIdProviderMetadata().getClientRegistrationTypesSupported().contains("explicit") || opStatement.getMetadata().getOpenIdProviderMetadata().getFederationRegistrationEndpoint() == null) {
-                    throw new BadRequestException("No valid OP Entity Statement");
-                }
-                TrustChainResolution trustChainResolution = trustChainProcessor.constructTrustChains(opStatement, Stream.of(federationConfig.getTrustAnchor()).collect(Collectors.toSet()),  false);
-                if (trustChainResolution == null) {
-                    throw new BadRequestException("No common trust chain found");
-                }
-                OPMetadata op = (OPMetadata) trustChainResolution.getMetadataAfterPolicies();
-                IdentityProviderModel model = OIDCIdentityProviderFactory.parseOIDCConfig(op,  OpenIdFederationIdentityProviderConfig.class, new OpenIdFederationIdentityProviderConfig());
-                if (representation.getConfig().get("guiOrder") != null && !representation.getConfig().get("guiOrder").isEmpty()) {
+                  IdentityProviderModel model = new OpenIdFederationIdentityProviderConfig();
+                  if (representation.getConfig().get("guiOrder") != null && !representation.getConfig().get("guiOrder").isEmpty()) {
                     model.getConfig().put("guiOrder", representation.getConfig().get("guiOrder"));
-                }
-
-                UriInfo frontendUriInfo = session.getContext().getUri(UrlType.FRONTEND);
-                UriInfo backendUriInfo = session.getContext().getUri(UrlType.BACKEND);
-                JSONWebKeySet jwks = trustChainProcessor.getKeySet();
-                EntityStatement entityStatement = new EntityStatement(Urls.realmIssuer(frontendUriInfo.getBaseUri(), realm.getName()), Long.valueOf(federationGeneralConfig.getLifespan()), Stream.of(trustChainResolution.getLeafId()).collect(Collectors.toList()), jwks);
-                entityStatement.addAudience(opIssuer);
-                Metadata metadata = new Metadata();
-                // Use RP client registration types from general config
-                Stream<ClientRegistrationTypeEnum> rpRegistrationTypes = federationGeneralConfig.getRpClientRegistrationTypesSupported() != null ? 
-                    federationGeneralConfig.getRpClientRegistrationTypesSupported().stream() : 
-                    Stream.empty();
-                RPMetadata rPMetadata = OpenIdFederationUtils.createRPMetadata(federationGeneralConfig, rpRegistrationTypes, OpenIdFederationUtils.commonMetadata(federationGeneralConfig), RealmsResource.protocolUrl(backendUriInfo).clone().path(OIDCLoginProtocolService.class, "certs").build(realm.getName(),
-                        OIDCLoginProtocol.LOGIN_PROTOCOL).toString(), frontendUriInfo, realm.getName());
-                metadataFromOP(rPMetadata, federationConfig.getIdpConfiguration(), op, opStatement.getSubject());
-                metadataFromFederation(rPMetadata, federationConfig.getIdpConfiguration());
-                metadata.setRelyingPartyMetadata(rPMetadata);
-                entityStatement.setMetadata(metadata);
-                StringEntity entity = new StringEntity(session.tokens().encodeForOpenIdFederation(entityStatement), TokenUtil.APPLICATION_ENTITY_STATEMENT_JWT);
-                SimpleHttpResponse response = SimpleHttp.create(session).doPost(op.getFederationRegistrationEndpoint())
-                        .header("Content-Type", "application/entity-statement+jwt")
-                        .entity(entity)
-                        .asResponse();
-                if (response.getStatus() < 200 || response.getStatus() >= 400) {
-                    throw new BadRequestException("Error during explicit client registration with body : "+ response.asString());
-                }
-                EntityStatementExplicitResponse statementResponse = trustChainProcessor.parseAndValidateSelfSigned(response.asString(), EntityStatementExplicitResponse.class, opStatement.getJwks());
-                if (!trustChainProcessor.validateEntityStatementFields(statementResponse, opIssuer, opIssuer) || statementResponse.getTrustAnchor() == null || LocalDateTime.now().toEpochSecond(ZoneOffset.UTC) > statementResponse.getExp() ) {
-                    throw new BadRequestException("No valid OP Entity Statement");
-                }
-                OpenIdFederationUtils.convertEntityStatementToIdp(model, realm, representation.getAlias(), statementResponse, new HashMap<>(federationConfig.getIdpConfiguration()));
-                return model;
+                  }
+                  model.setAlias(representation.getAlias());
+                  TrustChainProcessor trustChainProcessor = session.getProvider(TrustChainProcessor.class, OpenIdFederationTrustChainProcessorFactory.PROVIDER_ID);
+                  trustChainProcessor.rPexcplicitRegistration(representation.getConfig().get(OIDCIdentityProviderConfig.ISSUER), representation.getConfig().get(OpenIdFederationIdentityProviderConfig.TRUST_ANCHOR_ID), model, realm);
+                  return model;
             } catch (Exception e) {
                 throw ErrorResponse.error(e.getMessage(), BAD_REQUEST);
             }
@@ -368,22 +297,6 @@ public class IdentityProvidersResource {
         AutoUpdateIdentityProviders autoUpdateProvider = new AutoUpdateIdentityProviders(alias, realm.getId());
         ClusterAwareScheduledTaskRunner taskRunner = new ClusterAwareScheduledTaskRunner(session.getKeycloakSessionFactory(), autoUpdateProvider, interval);
         timer.schedule(taskRunner, interval, realm.getId()+"_AutoUpdateIdP_" + alias);
-    }
-
-    private void metadataFromFederation(RPMetadata rPMetadata, Map<String, String> federationConfig){
-        rPMetadata.setScope(federationConfig.get(OAuth2IdentityProviderConfig.DEFAULT_SCOPE));
-    }
-
-    private void metadataFromOP(RPMetadata rPMetadata, Map<String, String> federationConfig, OPMetadata opMetadata, String subject) {
-        List<String> subjectTypesSupported = federationConfig.get(OpenIdFederationUtils.SUBJECT_TYPES_SUPPORTED) == null
-                ? OIDCWellKnownProvider.DEFAULT_SUBJECT_TYPES_SUPPORTED
-                : Arrays.asList(federationConfig.get(OpenIdFederationUtils.SUBJECT_TYPES_SUPPORTED).split("##"));
-
-        rPMetadata.setSubjectType(subjectTypesSupported.stream()
-                .filter(x -> opMetadata.getSubjectTypesSupported().contains(x))
-                .findFirst()
-                .orElseThrow(() -> new BadRequestException("No subject type common exists")));
-        rPMetadata.setClientName(opMetadata.getCommonMetadata().getOrganizationName() != null ? opMetadata.getCommonMetadata().getOrganizationName() : subject);
     }
 
     @Path("instances/{alias}")