+++ services/src/main/java/org/keycloak/utils/OpenIdFederationTrustChainProcessorFactory.java
@@ -3,6 +3,8 @@ package org.keycloak.utils;
 import org.keycloak.Config;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.KeycloakSessionFactory;
+import org.keycloak.protocol.trustchain.TrustChainProcessor;
+import org.keycloak.protocol.trustchain.TrustChainProcessorFactory;
 
 public class OpenIdFederationTrustChainProcessorFactory implements TrustChainProcessorFactory {